Refactor useCache separate storage modules
